requirements.txt
aws-opentelemetry-distro>=0.10.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
litellm
mcp[cli]
nova-act
opensearch-py
pandas
retrying
strands-agents 
strands-agents-tools[mem0_memory]
streamlit
tqdm
uv
botocore
awscli